backend/requirements.txt
asgiref==3.11.1
attrs==26.1.0
Django==6.0.3
django-cors-headers==4.9.0
django-environ==0.13.0
django-filter==25.2
django-unfold==0.88.0
djangorestframework==3.17.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
pillow==12.2.0
python-decouple==3.8
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
tzdata==2026.1
uritemplate==4.2.0
whitenoise==6.12.0
gunicorn==23.0.0
dj-database-url==2.1.0
psycopg2-binary==2.9.10
